+++ precompiles/werc20/testdata/wevmos9.go
@@ -1,0 +1,15 @@
+// Copyright Tharsis Labs Ltd.(Evmos)
+// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
+
+package testdata
+
+import (
+	contractutils "github.com/evmos/evmos/v20/contracts/utils"
+	evmtypes "github.com/evmos/evmos/v20/x/evm/types"
+)
+
+// LoadWEVMOS9Contract load the WEVMOS9 contract from the json representation of
+// the Solidity contract.
+func LoadWEVMOS9Contract() (evmtypes.CompiledContract, error) {
+	return contractutils.LoadContractFromJSONFile("WEVMOS9.json")
+}