+++ CHANGELOG.md
@@ -64,7 +64,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (evm) [#2936](https://github.com/evmos/evmos/pull/2936) Add query for EVM config.
 - (app) [#2937](https://github.com/evmos/evmos/pull/2937) Fix conversion on the `CheckTxFee` ante handler and allow zero coins refunds.
 - (deps) [#2967](https://github.com/evmos/evmos/pull/2967) Bump CometBFT to `v0.38.15`.
-- (precompiles) [#2966](https://github.com/evmos/evmos/pull/2966) Add safety check that ERC20 precompiles cannot receive funds.
+- (precompiles) [#2943](https://github.com/evmos/evmos/pull/2943) Add WERC-20 precompile.
+- (precompiles) [#2966](https://github.com/evmos/evmos/pull/2966) Add safety check that ERC-20 precompiles cannot receive funds.
 
 ### Bug Fixes